- Standardize data types in numeric operations across Bézier and NURBS methods for consistency and precision. - Add `examples/ssx/nurbs_nurbs_intersection_10.py` to demonstrate NURBS intersection use cases. - Refactor `ndinterval.c` to fix method names and refine interval operations. - Introduce `_deflate.py` for advanced intersection processing with improved derivative handling and multi-dimensional Bernstein patch methods. - Optimize memory usage by caching frequently used computations in `fdm.py` and `aabb.py`.

+++ mmcore/numeric/bern.py
@@ -536,7 +536,7 @@ def bernstein_trim_nd(control_grid: np.ndarray, ranges) -> np.ndarray:
         raise ValueError("Need at least one parametric axis plus a trailing value axis.")
 
     D = control_grid.ndim - 1
-    ranges = np.asarray(ranges,dtype=ranges.dtype)
+    ranges = np.asarray(ranges,)
     if ranges.shape != (D, 2):
         raise ValueError(f"`ranges` must have shape ({D}, 2).")
